refactor: drop MLC prefix, merge utility/constants/enum files
Rename MLCUtils -> LocalizationUtils and fold MLCFsmUtils into the same
file as a second static class FsmUtils. Move LocalizationConstants into
LocalizationConfig.cs and the TranslationMode enum into TranslationPattern.cs.
Net: -3 files, no behavior change.
