+++ MWC_Localization_Core.csproj
@@ -70,7 +70,7 @@
     <Compile Include="CoreConsole.cs" />
     <Compile Include="LateUpdateHandler.cs" />
     <Compile Include="LocalizationConfig.cs" />
-    <Compile Include="LocalizationConstants.cs" />
+    <Compile Include="LocalizationUtils.cs" />
     <Compile Include="MagazineTextHandler.cs" />
     <Compile Include="FsmTextHook.cs" />
     <Compile Include="FsmTextHook.BuiltInTargets.cs" />
@@ -80,12 +80,9 @@
     <Compile Include="TextAdjustment.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SceneTranslationManager.cs" />
-    <Compile Include="MLCFsmUtils.cs" />
-    <Compile Include="MLCUtils.cs" />
     <Compile Include="TeletextHandler.cs" />
     <Compile Include="TextMeshTranslator.cs" />
     <Compile Include="TranslationFileParser.cs" />
-    <Compile Include="TranslationMode.cs" />
     <Compile Include="TranslationPattern.cs" />
     <Compile Include="UnifiedTextMeshMonitor.cs" />
   </ItemGroup>
