+++ TranslationPattern.cs
@@ -4,6 +4,27 @@ using System.Collections.Generic;
 namespace MWC_Localization_Core
 {
     /// <summary>
+    /// Defines which translation method to use.
+    /// </summary>
+    public enum TranslationMode
+    {
+        /// <summary>
+        /// Pattern matching with {0}, {1}, and similar placeholders.
+        /// </summary>
+        FsmPattern,
+
+        /// <summary>
+        /// Use a custom handler for more complex translation logic.
+        /// </summary>
+        CustomHandler,
+
+        /// <summary>
+        /// Pattern matching with placeholders plus translation of extracted parameters.
+        /// </summary>
+        FsmPatternWithTranslation
+    }
+
+    /// <summary>
     /// Represents a translation pattern with extraction and formatting logic
     /// Supports placeholders and custom handlers
     /// </summary>
@@ -184,7 +205,7 @@ namespace MWC_Localization_Core
                 
                 // Try to translate the parameter value
                 string translatedValue = originalValue;
-                string key = MLCUtils.FormatUpperKey(originalValue);
+                string key = LocalizationUtils.FormatUpperKey(originalValue);
                 
                 if (translations.TryGetValue(key, out string translation))
                 {
